feat(08-01): add emit.rs and fix TransformModule.order to u64
- emit_module wraps OXC Codegen for AST -> JS with optional source maps
- EmitOptions{source_maps: bool} controls map generation
- EmitResult{code, map} carries output
- OXC Codegen emits double-quoted strings by default (matches qwik-core)
- TransformModule.order changed u32->u64 per SPEC.md alignment
- 3 emit tests (with maps, without maps, double-quote assertion)
- All 82 tests pass (60 Phase 7 + 22 new parse/emit tests)

+++ crates/qwik-optimizer-oxc/src/types.rs
@@ -194,8 +194,9 @@ pub struct TransformModule {
 
     /// Sort order for deterministic output ordering.
     /// Main modules get order 0; segments get their extraction order.
+    /// u64 to match SPEC.md (supports large segment counts).
     #[serde(default)]
-    pub order: u32,
+    pub order: u64,
 }
 
 /// Metadata about an extracted segment (a lazy-loadable code fragment).
